requirements.txt
azure-search-documents==11.6.0b12
azure-identity
openai
aiohttp
ipykernel
requests
python-dotenv